feat: integrate copilot chat with deals tab
